refactor(password): refactor common functions from password libs
