+++ src/libs/array-common.js
@@ -1,0 +1,17 @@
+export function pickNRandom(arr, n) {
+  let result = new Array(n),
+    len = arr.length,
+    taken = new Array(len);
+  if (n > len)
+    throw new RangeError('getRandom: more elements taken than available');
+  while (n--) {
+    const x = Math.floor(Math.random() * len);
+    result[n] = arr[x in taken ? taken[x] : x];
+    taken[x] = --len in taken ? taken[len] : len;
+  }
+  return result;
+}
+
+export function pickOneRandom(arr) {
+  return arr[Math.floor(Math.random() * arr.length)];
+}
